+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Updated to 1.21.1
-- Marked as beta as origins isn't stable for 1.21.1, and code may break
+- Updated to latest origins version (1.13.0-pre.2)